feat: 修复类型错误

+++ vscing-model/src/main/java/com/vscing/model/dto/OrganizationListDto.java
@@ -17,6 +17,6 @@ public class OrganizationListDto {
     private Integer type;
 
     @Schema(description = "机构ID/机构父ID")
-    private Integer organizationId;
+    private Long organizationId;
 
 }
